55: extract `<RecordNewExample>`, `<SingleExample>`, `<EditExample>`, `<Search>` into the newly-created `frontend/src/features/examples/` directory
