+++ frontend/src/App.test.tsx
@@ -15,14 +15,7 @@ import { createMemoryHistory } from "history";
 import { Router, Route } from "react-router-dom";
 
 import { IState, INITIAL_STATE, rootReducer } from "./store";
-import {
-  NavigationBar,
-  Home,
-  About,
-  Account,
-  RecordNewExample,
-  SingleExample,
-} from "./App";
+import { NavigationBar, Home, About, Account } from "./App";
 import { Alerts } from "./features/alerts/Alerts";
 
 // 2
@@ -236,186 +229,6 @@ describe("<Account>", () => {
       expect(tableCellElement).toBeInTheDocument();
     }
   });
-});
-
-describe("<RecordNewExample>", () => {
-  test("renders the fields of a form for creating a new Example resource", () => {
-    const realStore = createStore(rootReducer);
-    render(
-      <Provider store={realStore}>
-        <RecordNewExample />
-      </Provider>
-    );
-
-    const legendElement = screen.getByText("[legend-tag: CREATE NEW EXAMPLE]");
-    expect(legendElement).toBeInTheDocument();
-
-    const sourceLanguageLabelElement = screen.getByText("SOURCE LANGUAGE");
-    expect(sourceLanguageLabelElement).toBeInTheDocument();
-
-    const newWordLabelElement = screen.getByText("NEW WORD");
-    expect(newWordLabelElement).toBeInTheDocument();
-
-    const exampleLabelElement = screen.getByText("EXAMPLE");
-    expect(exampleLabelElement).toBeInTheDocument();
-
-    const translationLabelElement = screen.getByText("TRANSLATION");
-    expect(translationLabelElement).toBeInTheDocument();
-
-    const submitInputElement = screen.getByRole("button", {
-      name: "RECORD THIS EXAMPLE",
-    });
-    expect(submitInputElement).toBeInTheDocument();
-  });
-
-  test(
-    "+ <Alerts>" +
-      " - renders an alert after the user has submitted the form" +
-      " without completing all required fields",
-    () => {
-      /* Arrange. */
-      const realStore = createStore(rootReducer);
-
-      render(
-        <Provider store={realStore}>
-          <Alerts />
-          <RecordNewExample />
-        </Provider>
-      );
-
-      const newWordInputElement = screen.getByLabelText("NEW WORD");
-      expect(newWordInputElement).toBeInTheDocument();
-      // const exampleInputElement = screen.getByLabelText("EXAMPLE");
-      // expect(exampleInputElement).toBeInTheDocument();
-
-      fireEvent.change(newWordInputElement, {
-        target: { value: "test-new-word" },
-      });
-      // fireEvent.change(exampleInputElement, {
-      //   target: { value: "test-example" },
-      // });
-
-      /* Act. */
-      const submitButtonElement = screen.getByRole("button", {
-        name: "RECORD THIS EXAMPLE",
-      });
-      fireEvent.click(submitButtonElement);
-
-      /* Assert. */
-      screen.getByText(
-        "YOU MUST FILL OUT THE FOLLOWING FORM FIELDS: NEW WORD, EXAMPLE"
-      );
-    }
-  );
-});
-
-describe("<SingleExample>", () => {
-  test(
-    "renders a specific Example resource" +
-      " and HTML elements that enable user interaction",
-    () => {
-      /* Arrange. */
-      const perPage: number = 2;
-      const page: number = 1;
-
-      const paginationFromBackend: {
-        _meta: IPaginationMetaFromBackend;
-        _links: IPaginationLinks;
-        items: IExampleFromBackend[];
-      } = mockPaginationFromBackend(examplesMock, perPage, page);
-      const {
-        meta,
-        links,
-        ids,
-        entities,
-      }: {
-        meta: IPaginationMeta;
-        links: IPaginationLinks;
-        ids: number[];
-        entities: { [exampleId: string]: IExample };
-      } = convertToPaginationInFrontend(paginationFromBackend);
-
-      const initState: IState = {
-        ...INITIAL_STATE,
-        examples: {
-          ...INITIAL_STATE.examples,
-          meta,
-          links,
-          ids,
-          entities,
-        },
-      };
-      const enhancer = applyMiddleware(thunkMiddleware);
-      const realStore = createStore(rootReducer, initState, enhancer);
-
-      const history = createMemoryHistory();
-      history.push("/example/2");
-
-      /* Arrange. */
-      render(
-        <Provider store={realStore}>
-          <Router history={history}>
-            <Route exact path="/example/:id">
-              <SingleExample />
-            </Route>
-          </Router>
-        </Provider>
-      );
-
-      /* Assert. */
-      const divElement = screen.getByText(
-        "You have selected the following Example from your Own VocabTreasury:"
-      );
-      expect(divElement).toBeInTheDocument();
-
-      /* First row. */
-      const idTableCellElement1 = screen.getByText("ID");
-      expect(idTableCellElement1).toBeInTheDocument();
-
-      const sourceLanguageTableCellElement1 =
-        screen.getByText("SOURCE LANGUAGE");
-      expect(sourceLanguageTableCellElement1).toBeInTheDocument();
-
-      const newWordTableCellElement1 = screen.getByText("NEW WORD");
-      expect(newWordTableCellElement1).toBeInTheDocument();
-
-      const exampleTableCellElement1 = screen.getByText("EXAMPLE");
-      expect(exampleTableCellElement1).toBeInTheDocument();
-
-      const translationTableCellElement1 = screen.getByText("TRANSLATION");
-      expect(translationTableCellElement1).toBeInTheDocument();
-
-      /* Second row. */
-      const idTableCellElement2 = screen.getByText("2");
-      expect(idTableCellElement2).toBeInTheDocument();
-
-      const sourceLanguageTableCellElement2 = screen.getByText("Finnish");
-      expect(sourceLanguageTableCellElement2).toBeInTheDocument();
-
-      const newWordTableCellElement2 = screen.getByText("sana numero-2");
-      expect(newWordTableCellElement2).toBeInTheDocument();
-
-      const exampleTableCellElement2 = screen.getByText("lause numero-2");
-      expect(exampleTableCellElement2).toBeInTheDocument();
-
-      const translationTableCellElement2 = screen.getByText("käännös numero-2");
-      expect(translationTableCellElement2).toBeInTheDocument();
-
-      // HTML elements that enable user interaction.
-      const anchorForReturning = screen.getByText(
-        "Return to this example within my Own VocabTreasury"
-      );
-      expect(anchorForReturning).toBeInTheDocument();
-
-      const anchorForEditing = screen.getByText("Edit this example");
-      expect(anchorForEditing).toBeInTheDocument();
-
-      const buttonForDeleting = screen.getByRole("button", {
-        name: "Delete this example",
-      });
-      expect(buttonForDeleting).toBeInTheDocument();
-    }
-  );
 });
 
 /* Describe what requests should be mocked. */
